requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==3.2
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-js-asset==2.2.0
django-mptt==0.15.0
django-phonenumber-field==7.2.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
phonenumbers==8.13.27
phonenumberslite==8.13.27
Pillow==10.1.0
psycopg2-binary==2.9.6
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
